# `exp`, optimizations, clean-up, bugfixes, v0.1.1