--- a/ci/script.sh
+++ b/ci/script.sh
@@ -3,10 +3,6 @@ set -ex
 
 cargo=cargo
 target_param=""
-features=""
-if [ ! -z "$UNSTABLE" ]; then
-    features+=" unstable"
-fi
 if [ ! -z "$TARGET" ]; then
     rustup target add "$TARGET"
     cargo install -v cross --force
@@ -14,15 +10,15 @@ if [ ! -z "$TARGET" ]; then
     target_param="--target $TARGET"
 fi
 
-$cargo build -v $target_param --features "$features"
-$cargo test -v $target_param --features "$features"
+$cargo build -v $target_param
+$cargo test -v $target_param
 
 # for now, `cross bench` is broken https://github.com/rust-embedded/cross/issues/239
 if [ "$cargo" != "cross" ]; then
-    $cargo bench -v $target_param --features "$features" -- --test # don't actually record numbers
+    $cargo bench -v $target_param -- --test # don't actually record numbers
 fi
 
-$cargo doc -v $target_param --features "$features"
+$cargo doc -v $target_param
 
 $cargo test -v --release
 
@@ -33,6 +29,6 @@ if [ ! -z "$COVERAGE" ]; then
     fi
 
     cargo install -v cargo-travis || echo "cargo-travis already installed"
-    cargo coverage -v -m coverage-reports --kcov-build-location "$PWD/target" --features "$features"
+    cargo coverage -v -m coverage-reports --kcov-build-location "$PWD/target"
     bash <(curl -s https://codecov.io/bash) -c -X gcov -X coveragepy -s coverage-reports
 fi