## Benchmarks

Prefix:

- `scalar/...` are benchmarks that are doing a single operation in a way that is hard/impossible to autovectorize,
- `vector/...` are doing it in a way that it is easy to autovectorize.

Suffix:

- `baseline` are doing nothing, to measure the overhead of just iterating,
- `full` are the complete `fast_math` versions that work for all inputs
- `raw` are the `..._raw` ones
- `std` are the full-accuracy functions from `std`

## `exp` performance
```
scalar/exp/baseline     time:   [5.4553 ns 5.4777 ns 5.5041 ns]
scalar/exp/full         time:   [35.651 ns 35.802 ns 35.972 ns]
scalar/exp/raw          time:   [37.982 ns 38.083 ns 38.174 ns]
scalar/exp/std          time:   [69.198 ns 69.527 ns 69.894 ns]

vector/exp/baseline     time:   [4.0552 ns 4.0728 ns 4.0912 ns]
vector/exp/full         time:   [15.189 ns 15.248 ns 15.311 ns]
vector/exp/raw          time:   [8.3843 ns 8.4221 ns 8.4641 ns]
vector/exp/std          time:   [77.019 ns 77.390 ns 77.807 ns]

scalar/exp2/baseline    time:   [5.4680 ns 5.4993 ns 5.5343 ns]
scalar/exp2/full        time:   [33.331 ns 33.686 ns 34.146 ns]
scalar/exp2/raw         time:   [37.700 ns 37.863 ns 38.044 ns]
scalar/exp2/std         time:   [66.424 ns 66.829 ns 67.352 ns]

vector/exp2/baseline    time:   [4.1093 ns 4.1430 ns 4.1874 ns]
vector/exp2/full        time:   [14.316 ns 14.358 ns 14.413 ns]
vector/exp2/raw         time:   [8.4198 ns 8.4637 ns 8.5158 ns]
vector/exp2/std         time:   [79.952 ns 80.444 ns 81.007 ns]
```
## Optimizations

The `change` lines are the comparison to the performance before any of the optimizations.

Summary:

- `log2`: ~20% less time overall (25% faster), but also possible to autovectorize (80% less time&mdash;5 &times; faster&mdash;on `vector/log2/raw`)
- `atan`: 70% less time overall (>3&times; faster), 80% less time when vectorizing
- `atan2`: 70% less time overall (>3&times; faster)

```
scalar/log2/baseline    time:   [5.4830 ns 5.5150 ns 5.5543 ns]
scalar/log2/full        time:   [51.118 ns 51.311 ns 51.497 ns]
                        change: [-20.418% -19.615% -18.938%] (p = 0.00 < 0.05)
scalar/log2/raw         time:   [38.746 ns 38.923 ns 39.119 ns]
                        change: [-24.566% -24.217% -23.875%] (p = 0.00 < 0.05)
scalar/log2/std         time:   [79.249 ns 79.569 ns 79.933 ns]

vector/log2/baseline    time:   [4.0939 ns 4.1389 ns 4.1896 ns]
vector/log2/full        time:   [48.569 ns 48.699 ns 48.846 ns]
                        change: [-24.008% -23.684% -23.350%] (p = 0.00 < 0.05)
vector/log2/raw         time:   [10.255 ns 10.379 ns 10.547 ns]
                        change: [-80.624% -80.492% -80.354%] (p = 0.00 < 0.05)
vector/log2/std         time:   [88.425 ns 88.755 ns 89.142 ns]
```

```
scalar/atan/baseline    time:   [5.4947 ns 5.5785 ns 5.6840 ns]
scalar/atan/full        time:   [24.910 ns 25.033 ns 25.159 ns]
                        change: [-74.214% -73.960% -73.757%] (p = 0.00 < 0.05)
scalar/atan/raw         time:   [13.184 ns 13.236 ns 13.287 ns]
                        change: [-63.833% -63.670% -63.505%] (p = 0.00 < 0.05)
scalar/atan/std         time:   [75.356 ns 75.871 ns 76.452 ns]

vector/atan/baseline    time:   [4.0432 ns 4.0581 ns 4.0752 ns]
vector/atan/full        time:   [11.622 ns 11.682 ns 11.745 ns]
                        change: [-87.538% -87.477% -87.418%] (p = 0.00 < 0.05)
vector/atan/raw         time:   [6.1044 ns 6.1332 ns 6.1627 ns]
                        change: [-83.680% -83.564% -83.457%] (p = 0.00 < 0.05)
vector/atan/std         time:   [76.743 ns 77.039 ns 77.361 ns]
```
```
scalar/atan2/baseline   time:   [7.1217 ns 7.1496 ns 7.1786 ns]
scalar/atan2/full       time:   [19.184 ns 19.262 ns 19.352 ns]
                        change: [-69.945% -69.663% -69.403%] (p = 0.00 < 0.05)
scalar/atan2/std        time:   [47.187 ns 47.402 ns 47.641 ns]
```